chore: Extract server code into smaller modules
