+++ backend/server/ assets/privacy-policy.md
