+++ backend/server/env.d.ts
@@ -1,0 +1,1 @@
+/// <reference types="@aklinker1/aframe/env" />
