+++ backend/server/plugins/models.ts
@@ -1,0 +1,16 @@
+import { Elysia } from "elysia";
+import {
+  AiModel,
+  AiModelEnum,
+  ChatMessage,
+  PostChatRequestBody,
+} from "../../shared/types";
+
+export const models = new Elysia({ name: "models" })
+  .model({
+    AiModel,
+    AiModelEnum,
+    ChatMessage,
+    PostChatRequestBody,
+  })
+  .as("plugin");
