+++ backend/server/plugins/resolve-ip.ts
@@ -1,0 +1,7 @@
+import { Elysia } from "elysia";
+
+export const resolveIp = new Elysia({ name: "resolve-ip" })
+  .resolve(() => {
+    // TODO
+  })
+  .as("plugin");
