+++ backend/shared/constants.ts
@@ -1,0 +1,1 @@
+export { version } from "../package.json";
