+++ backend/tsconfig.json
@@ -1,4 +1,4 @@
 {
   "extends": "../tsconfig.base.json",
-  "include": ["./src"]
+  "include": ["./app", "./server", "./shared"]
 }
