+++ package.json
@@ -1,10 +1,13 @@
 {
+  "devDependencies": {
+    "@aklinker1/check": "^1.4.5",
+    "@types/bun": "latest",
+    "npm-run-all": "^4.1.5",
+    "oxlint": "^0.15.5",
+    "prettier": "^3.4.2",
+    "typescript": "^5.0.0"
+  },
   "packageManager": "bun@1.2.3",
-  "type": "module",
-  "workspaces": [
-    "plugin",
-    "backend"
-  ],
   "scripts": {
     "dev": "run-p dev:*",
     "dev:docs": "vitepress dev docs",
@@ -14,12 +17,9 @@
     "docs:build": "vitepress build docs",
     "docs:preview": "vitepress preview docs"
   },
-  "devDependencies": {
-    "@aklinker1/check": "^1.4.5",
-    "@types/bun": "latest",
-    "npm-run-all": "^4.1.5",
-    "oxlint": "^0.15.5",
-    "prettier": "^3.4.2",
-    "typescript": "^5.0.0"
-  }
+  "type": "module",
+  "workspaces": [
+    "plugin",
+    "backend"
+  ]
 }
